+++ src/Uwave.js
@@ -3,7 +3,7 @@
 const EventEmitter = require('events');
 const { promisify } = require('util');
 const mongoose = require('mongoose');
-const Redis = require('ioredis');
+const Redis = require('ioredis').default;
 const debug = require('debug');
 const avvio = require('avvio');
 
@@ -36,7 +36,7 @@ const DEFAULT_REDIS_URL = 'redis://localhost:6379';
 /**
  * @typedef {UwaveServer & avvio.Server<UwaveServer>} Boot
  * @typedef {Pick<
- *   Redis.RedisOptions,
+ *   import('ioredis').RedisOptions,
  *   'port' | 'host' | 'family' | 'path' | 'db' | 'password' | 'username' | 'tls'
  * >} RedisOptions
  * @typedef {{
@@ -47,75 +47,75 @@ const DEFAULT_REDIS_URL = 'redis://localhost:6379';
  */
 
 class UwaveServer extends EventEmitter {
-  /** @type {Redis.Redis} */
+  /** @type {import('ioredis').default} */
   redis;
 
   /** @type {import('http').Server} */
-  // @ts-ignore TS2564 Definitely assigned in a plugin
+  // @ts-expect-error TS2564 Definitely assigned in a plugin
   server;
 
   /** @type {import('express').Application} */
-  // @ts-ignore TS2564 Definitely assigned in a plugin
+  // @ts-expect-error TS2564 Definitely assigned in a plugin
   express;
 
   /** @type {import('./models').Models} */
-  // @ts-ignore TS2564 Definitely assigned in a plugin
+  // @ts-expect-error TS2564 Definitely assigned in a plugin
   models;
 
   /** @type {import('./plugins/acl').Acl} */
-  // @ts-ignore TS2564 Definitely assigned in a plugin
+  // @ts-expect-error TS2564 Definitely assigned in a plugin
   acl;
 
   /** @type {import('./plugins/bans').Bans} */
-  // @ts-ignore TS2564 Definitely assigned in a plugin
+  // @ts-expect-error TS2564 Definitely assigned in a plugin
   bans;
 
   /** @type {import('./plugins/booth').Booth} */
-  // @ts-ignore TS2564 Definitely assigned in a plugin
+  // @ts-expect-error TS2564 Definitely assigned in a plugin
   booth;
 
   /** @type {import('./plugins/chat').Chat} */
-  // @ts-ignore TS2564 Definitely assigned in a plugin
+  // @ts-expect-error TS2564 Definitely assigned in a plugin
   chat;
 
   /** @type {import('./plugins/configStore').ConfigStore} */
-  // @ts-ignore TS2564 Definitely assigned in a plugin
+  // @ts-expect-error TS2564 Definitely assigned in a plugin
   config;
 
   /** @type {import('./plugins/history').HistoryRepository} */
-  // @ts-ignore TS2564 Definitely assigned in a plugin
+  // @ts-expect-error TS2564 Definitely assigned in a plugin
   history;
 
   /** @type {import('./plugins/migrations').Migrate} */
-  // @ts-ignore TS2564 Definitely assigned in a plugin
+  // @ts-expect-error TS2564 Definitely assigned in a plugin
   migrate;
 
   /** @type {import('./plugins/motd').MOTD} */
-  // @ts-ignore TS2564 Definitely assigned in a plugin
+  // @ts-expect-error TS2564 Definitely assigned in a plugin
   motd;
 
   /** @type {import('./plugins/passport').Passport} */
-  // @ts-ignore TS2564 Definitely assigned in a plugin
+  // @ts-expect-error TS2564 Definitely assigned in a plugin
   passport;
 
   /** @type {import('./plugins/playlists').PlaylistsRepository} */
-  // @ts-ignore TS2564 Definitely assigned in a plugin
+  // @ts-expect-error TS2564 Definitely assigned in a plugin
   playlists;
 
   /** @type {import('./plugins/users').UsersRepository} */
-  // @ts-ignore TS2564 Definitely assigned in a plugin
+  // @ts-expect-error TS2564 Definitely assigned in a plugin
   users;
 
   /** @type {import('./plugins/waitlist').Waitlist} */
-  // @ts-ignore TS2564 Definitely assigned in a plugin
+  // @ts-expect-error TS2564 Definitely assigned in a plugin
   waitlist;
 
   /** @type {import('./HttpApi').HttpApi} */
-  // @ts-ignore TS2564 Definitely assigned in a plugin
+  // @ts-expect-error TS2564 Definitely assigned in a plugin
   httpApi;
 
   /** @type {import('./SocketServer')} */
-  // @ts-ignore TS2564 Definitely assigned in a plugin
+  // @ts-expect-error TS2564 Definitely assigned in a plugin
   socketServer;
 
   /**
@@ -309,8 +309,8 @@ class UwaveServer extends EventEmitter {
 
   async listen() {
     /** @type {import('avvio').Avvio<this>} */
-    // @ts-ignore
-    const boot = this;
+    // @ts-expect-error TS2322
+    const boot = this; // eslint-disable-line @typescript-eslint/no-this-alias
     await boot.ready();
 
     /** @type {(this: import('net').Server, port?: number) => Promise<void>} */