+++ src/plugins/history.js
@@ -51,9 +51,7 @@ class HistoryRepository {
       .populate('media.media user');
 
     /** @type {PopulatedHistoryEntry[]} */
-    // @ts-ignore TS2322: `media` and `user` types mismatch because typescript
-    // doesn't understand `populate()`.
-    const results = await query;
+    const results = /** @type {any} */ (await query);
 
     return new Page(results, {
       pageSize: pagination ? pagination.limit : undefined,